refactor: move date range changer into parent
